Projects to integrate/Caption Animator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "caption-animator"
version = "0.2.1"
description = "Render SRT/ASS subtitle files into animated, alpha-channel video overlays with plugin-based animation system."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Tim Eck", email = "public@verbumcaro.com" }
]

keywords = [
  "subtitles",
  "captions",
  "srt",
  "ass",
  "ffmpeg",
  "davinci-resolve",
  "video",
  "overlay",
  "animation",
  "plugins"
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Multimedia :: Video",
  "Topic :: Multimedia :: Sound/Audio",
  "Typing :: Typed"
]


dependencies = [
  "pysubs2>=1.6.0",
  "Pillow>=10.0.0",
  "PyYAML>=6.0"
]


[project.optional-dependencies]
dev = [
  "black",
  "ruff",
  "mypy",
  "pytest",
  "pytest-cov"
]


[project.scripts]
caption-animator = "caption_animator.cli.main:main"


[project.urls]
Homepage = "https://github.com/pulsence/caption-animator"
Issues = "https://github.com/pulsence/caption-animator/issues"
Documentation = "https://github.com/pulsence/caption-animator#readme"


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


[tool.black]
line-length = 100
target-version = ["py39"]


[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]
src = ["src"]


[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
